// src/types/transaction.ts
export enum type {
  DEPOSIT = "DEPOSIT",
  TRANSFER = "TRANSFER",
  EXTERNAL_PAYMENT = "EXTERNAL_PAYMENT",
  WITHDRAWAL = "WITHDRAWAL",
  REFUND = "REFUND",
  OTHER = "OTHER",
}

export enum transactionState {
  CREATED = "CREATED",
  PROCESSING = "PROCESSING",
  SENT = "SENT",
  EXPIRED = "EXPIRED",
  DECLINED = "DECLINED",
  SUSPENDED = "SUSPENDED",
  REFUNDED = "REFUNDED",
  SUCCESSFUL = "SUCCESSFUL",
  REVERSED = "REVERSED",
}

interface AmountDetails {
  transactionAmount: number;
  transactionCurrency: Currency;
  country: Country;
}

interface DeviceData {
  batteryLevel: number;
  deviceLatitude: number;
  deviceLongitude: number;
  ipAddress: string;
  deviceIdentifier: string;
  vpnUsed: boolean;
  operatingSystem: string;
  deviceMaker: string;
  deviceModel: string;
  deviceYear: string;
  appVersion: string;
}

interface Tag {
  key: string;
  value: string;
}

export interface Transaction extends Document {
  type: type;
  transactionId: string;
  timestamp: number;
  originUserId: string;
  destinationUserId: string;
  transactionState: transactionState;
  originAmountDetails: AmountDetails;
  destinationAmountDetails: AmountDetails;
  promotionCodeUsed: boolean;
  reference: string;
  originDeviceData: DeviceData;
  destinationDeviceData: DeviceData;
  tags: Tag[];
  description: string;
}